# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.BulkUpdateStatusRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/models.TaskStatus'
        enum:
        - pending
        - in_progress
        - completed
      task_ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - status
    - task_ids
    type: object
...
  models.CreateTaskRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      due_date:
        type: string
      priority:
        allOf:
        - $ref: '#/definitions/models.TaskPriority'
        enum:
        - low
        - medium
        - high
      status:
        allOf:
        - $ref: '#/definitions/models.TaskStatus'
        enum:
        - pending
        - in_progress
        - completed
      title:
        maxLength: 200
        type: string
    required:
    - title
    type: object
...
  models.TaskPriority:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - TaskPriorityLow
    - TaskPriorityMedium
    - TaskPriorityHigh
...
  models.TaskStatus:
    enum:
    - pending
    - in_progress
    - completed
    type: string
    x-enum-varnames:
    - TaskStatusPending
    - TaskStatusInProgress
    - TaskStatusCompleted
...
  models.UpdateTaskRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      due_date:
        type: string
      priority:
        allOf:
        - $ref: '#/definitions/models.TaskPriority'
        enum:
        - low
        - medium
        - high
      status:
        allOf:
        - $ref: '#/definitions/models.TaskStatus'
        enum:
        - pending
        - in_progress
        - completed
      title:
        maxLength: 200
        type: string
    type: object
  models.UpdateTaskStatusRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/models.TaskStatus'
        enum:
        - pending
        - in_progress
        - completed
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@taskmanagement.com
    name: API Support
  description: RESTful API for managing tasks and categories
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Task Management API
  version: "1.0"
paths:
...
  /tasks:
    get:
      consumes:
      - application/json
      description: Get all tasks for the authenticated user with filtering, sorting,
        and pagination
      parameters:
      - description: Filter by status (pending, in_progress, completed)
        in: query
        name: status
        type: string
      - description: Filter by priority (low, medium, high)
        in: query
        name: priority
        type: string
      - description: Filter by category ID
        in: query
        name: category_id
        type: integer
      - description: Search in title and description
        in: query
        name: search
        type: string
      - default: created_at
        description: Sort by field (created_at, updated_at, due_date, priority)
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order (asc, desc)
        in: query
        name: sort_order
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size (max 100)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Tasks retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task for the authenticated user
      parameters:
      - description: Task creation data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Task created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Task deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid task ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete task
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Get a specific task by ID for the authenticated user
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Task retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid task ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get task by ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update an existing task (partial update)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task update data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Task updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update task
      tags:
      - tasks
  /tasks/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update only the status of a task (quick status change)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status update data
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTaskStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Task status updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update task status
      tags:
      - tasks
  /tasks/bulk/status:
    patch:
      consumes:
      - application/json
      description: Update status for multiple tasks at once
      parameters:
      - description: Bulk status update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BulkUpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Bulk update completed
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Bulk update task status
      tags:
      - tasks
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"